% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_data1.R
\docType{data}
\name{mlm_data1}
\alias{mlm_data1}
\title{A Simulated Multilevel Data}
\format{
A tibble with 5000 rows and 5 variables:
\describe{
\item{\code{cluster}}{Cluster ID}
\item{\code{id}}{Individual ID}
\item{\code{X1c}}{A continuous covariate}
\item{\code{X2b}}{A binary covariate}
\item{\code{Yij}}{Response Variable}
}
}
\usage{
mlm_data1
}
\description{
This data contains data of 100 clusters where each cluster contains 50 individuals.
}
\keyword{datasets}
