% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_data2.R
\docType{data}
\name{mlm_data2}
\alias{mlm_data2}
\title{A Simulated Multilevel Data}
\format{
A tibble with 10,500 rows and 6 variables:
\describe{
\item{\code{ea}}{Enumeration Area ID}
\item{\code{hh}}{Household ID}
\item{\code{id}}{Individual ID}
\item{\code{X1c}}{A continuous covariate}
\item{\code{X2b}}{A binary covariate}
\item{\code{Yijk}}{Response Variable}
}
}
\usage{
mlm_data2
}
\description{
This data contains data of 70 Enumeration Areas (EA) where each EA contains 30 Households (HH) and
each HH has 5 individuals.
}
\keyword{datasets}
