% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mor.R
\name{mor}
\alias{mor}
\title{Calculate Median Odds Ratio from a fitted multilevel binary logistic model object.}
\usage{
mor(object, se = TRUE, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{object}{An \code{glmerMod} object created by \code{\link[lme4:glmer]{lme4::glmer()}} or an \code{MixMod} object
created by \code{\link[GLMMadaptive:mixed_model]{GLMMadaptive::mixed_model()}} or
an \code{glmmTMB} object created by \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}}. See Details.}

\item{se}{Logical indicating whether or not to include the standard error of
MOR estimate. Defaults to \code{TRUE}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{TRUE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Currently not used.}
}
\value{
a \link[tibble:tibble-package]{tibble} with columns,
\itemize{
\item \code{term}:      Name of the estimate
\item \code{estimate}:  Estimate of Median Odds Ratio (MOR)
\item \code{std.error}: Standard Error of MOR
\item \code{ci_lower}:  Lower bound of the confidence interval of MOR
\item \code{ci_upper}:  Upper bound of the confidence interval of MOR
}
}
\description{
Calculate Median Odds Ratio from a fitted multilevel binary logistic model object.
}
\details{
Median Odds Ratio (MOR) is suggested as a measure of heterogeneity that quantifies
cluster heterogeneity and facilitates a direct comparison between covariate effects
and the magnitude of heterogeneity in terms of well-known odds ratios
\insertCite{larsen2000interpreting,larsen2005appropriate}{MOR}
}
\examples{
data("mlm_data1")
data("mlm_data2")

# fitting two level random intercept model using lme4 package
model <- lme4::glmer(Yij ~ X1c + X2b +  (1 | cluster),
              family = "binomial", data = mlm_data1)

mor(model)

# fitting two level random intercept model using GLMMadaptive package
model1 <- GLMMadaptive::mixed_model(fixed = Yij ~ X1c + X2b,
                                    random =  ~ 1 | cluster,
                                    family = binomial("logit"),
                                    data = mlm_data1)
mor(model1)

# fitting two level random intercept model using glmmTMB package
model2 <- glmmTMB::glmmTMB(Yij ~ X1c + X2b + (1 | cluster),
                           family = binomial("logit"), data = mlm_data1)

mor(model2)

# fitting three level random intercept model using glmmTMB package
model3 = glmmTMB::glmmTMB(Yijk ~ X1c + X2b + (1 | ea) + (1 | ea:hh),
                          family = "binomial", data = mlm_data2)

mor(model3)

# fitting three level random intercept model using lme4 package
model4 = lme4::glmer(Yijk ~ X1c + X2b + (1 | ea) + (1 | ea:hh),
                          family = "binomial", data = mlm_data2)


mor(model4)

}
\references{
\insertAllCited{}
}
