% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOR-package.R
\docType{package}
\name{MOR-package}
\alias{MOR}
\alias{MOR-package}
\title{MOR: Calculate Median Odds Ratio From a Fitted Multilevel Binary Logistic Model}
\description{
Provides a postestimation function to calculate the Median Odds Ratio (MOR) from a multilevel binary logistic regression model fit. Currently this package provides the Median Odds Ratio only in case of two-level random intercept model and three-level random intercept model where the model is fitted using the R packages 'lme4', 'GLMMadaptive' or 'glmmTMB'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shafayetShafee/MOR}
  \item \url{https://shafayetshafee.github.io/MOR/}
  \item Report bugs at \url{https://github.com/shafayetShafee/MOR/issues}
}

}
\author{
\strong{Maintainer}: Shafayet Khan Shafee \email{sshafee@isrt.ac.bd} (\href{https://orcid.org/0009-0002-8021-3788}{ORCID})

}
\keyword{internal}
