% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mor.R
\name{mor}
\alias{mor}
\title{Calculate Median Odds Ratio from a fitted multilevel binary logistic model object}
\usage{
mor(object, se = TRUE, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{object}{An \code{MixMod} object created by \code{GLMMadaptive::mixed_model()} or
an \code{glmmTMB} object created by \code{glmmTMB::glmmTMB()}.}

\item{se}{Logical indicating whether or not to include the standard error of
MOR estimate. Defaults to \code{TRUE}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{TRUE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{Currently not used.}
}
\value{
A named vector with elements:
\itemize{
\item \code{mor}:      Estimate of Median Odds Ratio
\item \code{se_mor}:   Standard Error of MOR
\item \code{ci_lower}: Lower bound on the confidence interval of MOR
\item \code{ci_upper}: Upper bound on the confidence interval of MOR
}
}
\description{
Calculate Median Odds Ratio from a fitted multilevel binary logistic model object
}
\examples{
library(GLMMadaptive)
data("guImmun", package = "mlmRev")
model <- mixed_model(fixed = immun ~ kid2p + mom25p + ord, random =  ~ 1 | mom,
                     family = binomial("logit"),
                     data = guImmun)

mor(model)

}
